# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Khurana Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface BookingInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone: string;
  serviceType: string;
  preferredDate: string;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @minLength 10 */
  message: string;
  /** @nullable */
  venue?: string | null;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusUpdate {
  status?: BookingStatusUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  name: string;
  email: string;
  phone: string;
  serviceType: string;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  budget?: string | null;
  preferredDate: string;
  message: string;
  /** @nullable */
  venue?: string | null;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface Service {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  category: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  startingPrice?: string | null;
  sortOrder: number;
}

export interface PortfolioItem {
  id: number;
  title: string;
  category: string;
  imageUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  description?: string | null;
  featured: boolean;
  /** @nullable */
  aspectRatio?: string | null;
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  clientName: string;
  /** @nullable */
  clientLocation?: string | null;
  /** @nullable */
  clientPhotoUrl?: string | null;
  review: string;
  rating: number;
  serviceType: string;
  featured: boolean;
  createdAt?: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  category: string;
  imageUrl: string;
  author?: string;
  readTime: string;
  publishedAt: string;
}

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  /** @minLength 10 */
  message: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface StudioStats {
  yearsExperience: number;
  happyClients: number;
  photosDelivered: number;
  clientRating: number;
  totalBookings: number;
  pendingBookings: number;
}

export type ListBookingsParams = {
status?: ListBookingsStatus;
serviceType?: string;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type ListPortfolioParams = {
category?: ListPortfolioCategory;
featured?: boolean;
};

export type ListPortfolioCategory = typeof ListPortfolioCategory[keyof typeof ListPortfolioCategory];


export const ListPortfolioCategory = {
  wedding: 'wedding',
  portrait: 'portrait',
  fashion: 'fashion',
  events: 'events',
  commercial: 'commercial',
  products: 'products',
  landscape: 'landscape',
  all: 'all',
} as const;

export type ListTestimonialsParams = {
featured?: boolean;
};

export type ListBlogPostsParams = {
category?: string;
limit?: number;
};