# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Khurana Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  category: string;
  imageUrl: string;
  author?: string;
  readTime: string;
  publishedAt: Date;
}