# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Khurana Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  name: string;
  email: string;
  phone: string;
  serviceType: string;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  budget?: string | null;
  preferredDate: string;
  message: string;
  /** @nullable */
  venue?: string | null;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}