# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Khurana Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone: string;
  serviceType: string;
  preferredDate: string;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  budget?: string | null;
  /** @minLength 10 */
  message: string;
  /** @nullable */
  venue?: string | null;
}