# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Khurana Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  category: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  startingPrice?: string | null;
  sortOrder: number;
}