# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Khurana Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  clientName: string;
  /** @nullable */
  clientLocation?: string | null;
  /** @nullable */
  clientPhotoUrl?: string | null;
  review: string;
  rating: number;
  serviceType: string;
  featured: boolean;
  createdAt?: Date;
}